Correct argument to int* RowStatus in CoinGetSolutionBasis() in trunk 1.5

+++ CoinMP/src/CoinMP.h
@@ -392,7 +392,7 @@ int    (SOLVCALL *CoinGetSolutionValues)(HPROB hProb, double* Activity, double* 
 													  double* SlackValues, double* ShadowPrice);
 int    (SOLVCALL *CoinGetSolutionRanges)(HPROB hProb, double* ObjLoRange, double* ObjUpRange,
 										           double* RhsLoRange, double* RhsUpRange);
-int    (SOLVCALL *CoinGetSolutionBasis)(HPROB hProb, int* ColStatus, double* RowStatus);
+int    (SOLVCALL *CoinGetSolutionBasis)(HPROB hProb, int* ColStatus, int* RowStatus);
 
 
 void   (SOLVCALL *CoinReadFile)(HPROB hProb, int FileType, const char* ReadFilename);
